requirements.txt
django
djangorestframework
elasticsearch
aiohttp
python-dotenv